Find plugins by unprefixed name
