+++ src/__test__/index.test.ts
@@ -7,17 +7,6 @@ import * as SwupTS from '../Swup.js';
 
 const baseUrl = window.location.origin;
 
-function createPlugin(plugin = {}) {
-	return {
-		name: 'TestPlugin',
-		isSwupPlugin: true as const,
-		mount: vi.fn(() => {}),
-		unmount: vi.fn(() => {}),
-		_checkRequirements: vi.fn(() => true),
-		...plugin
-	};
-}
-
 describe('Exports', () => {
 	it('should export Swup and Options/Plugin types', () => {
 		class SwupPlugin implements Plugin {
@@ -90,63 +79,5 @@ describe('ignoreVisit', () => {
 		swup.navigate('/path/');
 
 		expect(ignoreVisit.mock.calls).toHaveLength(1);
-	});
-});
-
-describe('Plugin module', () => {
-	it('should mount and unmount plugins', function () {
-		const plugin = createPlugin();
-		const swup = new Swup();
-		swup.use(plugin);
-		swup.unuse(plugin);
-
-		expect(plugin.mount.mock.calls).toHaveLength(1);
-		expect(plugin.unmount.mock.calls).toHaveLength(1);
-	});
-
-	it('should mount plugins from options', function () {
-		const plugin = createPlugin();
-		const swup = new Swup({ plugins: [plugin] });
-		expect(plugin.mount.mock.calls).toHaveLength(1);
-	});
-
-	it('should find a plugin instance by reference', function () {
-		const plugin = createPlugin({ name: 'ExamplePlugin' });
-		const swup = new Swup({ plugins: [plugin] });
-		const instance = swup.findPlugin(plugin);
-
-		expect(instance).toEqual(expect.objectContaining({ name: 'ExamplePlugin' }));
-	});
-
-	it('should find a plugin instance by name', function () {
-		const plugin = createPlugin({ name: 'ExamplePlugin' });
-		const swup = new Swup({ plugins: [plugin] });
-		const instance = swup.findPlugin('ExamplePlugin');
-
-		expect(instance).toEqual(expect.objectContaining({ name: 'ExamplePlugin' }));
-	});
-
-	it('should check plugin requirements', function () {
-		const plugin = createPlugin();
-		const swup = new Swup({ plugins: [plugin] });
-		expect(plugin._checkRequirements.mock.calls).toHaveLength(1);
-	});
-
-	it('should reject plugins with unmet requirements', function () {
-		const allowedPlugin = createPlugin({
-			name: 'AllowedPlugin',
-			_checkRequirements: () => true
-		});
-		const unallowedPlugin = createPlugin({
-			name: 'UnallowedPlugin',
-			_checkRequirements: () => false
-		});
-		const swup = new Swup({ plugins: [allowedPlugin, unallowedPlugin] });
-
-		const allowedInstance = swup.findPlugin(allowedPlugin);
-		expect(allowedInstance).toEqual(expect.objectContaining({ name: 'AllowedPlugin' }));
-
-		const unallowedInstance = swup.findPlugin(unallowedPlugin);
-		expect(unallowedInstance).toBeUndefined();
 	});
 });
